requirements.txt
elasticsearch>=8.19.1
pandas
google-cloud-aiplatform
kaggle
tqdm
a2a-sdk>=0.3.0
google-genai>=1.9.0
google-adk>=1.7.0
python-dotenv